异常记录bug修复

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesExceptionRecord.java
@@ -78,7 +78,7 @@ public class IcesExceptionRecord extends BaseEntity {
     /**
      * 影响因子
      */
-    private Long exrImpactFactor;
+    private Float exrImpactFactor;
     /**
      * 影响等级
      */
